requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.8.30
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
Django==5.1.4
django-channels==0.7.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
